refact: 2d and 3d folders
