+++ source/engine/draw.c
@@ -7,77 +7,3 @@ void	pixel_put(t_img *data, int x, int y, int color)
 	dst = data->addr + (y * data->line_length + x * (data->bits_per_pixel / 8));
 	*(unsigned int *)dst = color;
 }
-
-void	draw_line(t_img *img, int i[2], int f[2], int color)
-{
-	int	dist_x;
-	int	dist_y;
-
-	dist_x = ft_abs(f[0] - i[0]);
-	dist_y = ft_abs(f[1] - i[1]);
-	if (dist_x >= dist_y)
-	{
-		if (i[0] < f[0])
-			bresenham_x(img, i, f, color);
-		else
-			bresenham_x(img, f, i, color);
-	}
-	else
-	{
-		if (i[1] < f[1])
-			bresenham_y(img, i, f, color);
-		else
-			bresenham_y(img, f, i, color);
-	}
-}
-
-// void	draw_vertical_line(t_img *img, int x, int y0, int yf, int color)
-// {
-// 	while (y0 <= yf)
-// 	{
-// 		pixel_put(img, x, y0, color);
-// 		y0++;
-// 	}
-// }
-
-void	draw_background(t_data dat, int ceiling_color, int floor_color)
-{
-	int	x;
-	int	y;
-	int	color;
-
-	x = 0;
-	while (x < LENGHT)
-	{
-		y = 0;
-		color = ceiling_color;
-		while (y < HEIGHT)
-		{
-			if (y == HEIGHT / 2)
-				color = floor_color;
-			pixel_put(&dat.img, x, y, color);
-			y++;
-		}
-		x++;
-	}
-}
-
-void	draw_square(t_img *img, int x0, int y0, int color)
-{
-	int	x;
-	int	y;
-	int	size;
-
-	size = 22;
-	x = x0;
-	while (x < x0 + size)
-	{
-		y = y0;
-		while (y < y0 + size)
-		{
-			pixel_put(img, x, y, color);
-			y++;
-		}
-		x++;
-	}
-}
